tugas-3/requirements.txt
aiohttp==3.9.5
python-dotenv==1.0.1
redis==5.0.4
pytest==8.2.1
pytest-asyncio==0.23.7
locust==2.27.0
prometheus-client==0.20.0
